requirements.txt
Flask
flask_wtf
wtforms
flask-sqlalchemy
requests
python-dotenv